Bug fix; regex catching all eslint error
This error is caused when the regex is catching all errors returned by
eslint instead of catching only the relevant one.  This is because when
there are no intersecting lines when considering eslint and git diff output,
the initial part of regex goes empty instead of line numbers separated with
the pipe symbol.  So this regex will fetch all the errors instead of the errors
in git diff.